tspend: Support deterministic OP_RETURN
This adds a new argument '--deterministic' which forces generating a
deterministic OP_RETURN based on the payloads specified to the tool.

This reduces the chances of generating exactly duplicated tspends by
mistakenly calling the tool twice with the same set of inputs and
parameters (expiry, etc).

It also adds a log line to indicate the tspend was published or a
duplicated tspend was detected.